+++ client/src/Components/Common/index.tsx
@@ -1,0 +1,7 @@
+export { default as Alert } from "./Alert";
+export { default as Button } from "./Button";
+export { default as Checkbox } from "./Checkbox";
+export { default as Input } from "./Input";
+export { default as NoData } from "./NoData";
+export { default as NumInput } from "./NumInput";
+export { default as Rating } from "./Rating";
