+++ client/src/Components/ProductList/InfiniteScroll/index.tsx
@@ -5,7 +5,7 @@ import { media } from "@/styles/theme";
 import { Loading } from "@/shared/styled";
 import { useState } from "react";
 import { useEffect } from "react";
-import NoData from "@/Components/NoData";
+import NoData from "@/Components/Common/NoData";
 import { useLazyLoad } from "@/hooks/useLazyLoad";
 import { useRecoilValue } from "recoil";
 import { LocaitionStateType, locationState } from "@/store/history";
