+++ client/src/Pages/Main/BannerSection/index.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { Banner1, Banner2, Banner3 } from "@/assets";
 import Banner from "./Banner";
 import { ItemBannerType } from "@/shared/type";
-import Carousel from "@/Components/Carousel";
+import Carousel from "@/Pages/Main/BannerSection/Carousel";
 
 const banners: ItemBannerType[] = [
   {
