+++ client/src/Pages/Order/OrderContent/LoggedinContent.tsx
@@ -3,9 +3,9 @@ import useInput from "@/hooks/useInput";
 import AddressModal from "../AddressModal";
 import ItemInfoBox from "@/Components/ItemInfoBox";
 import { Arrow, ETPay, KakaoPay } from "@/assets";
-import InputSection from "@/Components/Input/InputSection";
+import InputSection from "@/Components/Common/Input/InputSection";
 import useValidation from "@/hooks/useValidation";
-import ValidationInput from "@/Components/Input/ValidationInput";
+import ValidationInput from "@/Components/Common/Input/ValidationInput";
 import { validatePhoneNumber, VALIDATION_ERR_MSG } from "@/utils/validations";
 import { useMyDestinations, useMyInfo } from "@/api/my";
 import { DestinationType, ICart, PartialCart } from "@/shared/type";
